Create PFreDA_models
Contains the data for models that can be used with the /vtem command to act as a sub model for GPT-4x to use as a basis for behavior and comparison. Models are extracted using PFreDA (phonetic frequency distribution analysis). /vtem and other function calls which can be stored using a paid ChatGPT.com account are stored in a separate file, titled, "ChatGPT_functions"